Extract WebSocketStreamingTest
