+++ core/src/test/scalajvm/sttp/client/testing/WebSocketStreamingTest.scala
@@ -1,0 +1,33 @@
+package sttp.client.testing
+
+import org.scalatest.Suite
+import org.scalatest.flatspec.AsyncFlatSpecLike
+import sttp.client.{asWebSocketStreamAlways, basicRequest}
+import sttp.client.testing.HttpTest.wsEndpoint
+import sttp.model.ws.WebSocketFrame
+import sttp.client._
+import sttp.client.monad._
+import sttp.client.monad.syntax._
+
+trait WebSocketStreamingTest[F[_], S] extends ToFutureWrapper { outer: Suite with AsyncFlatSpecLike =>
+  val streams: Streams[S]
+  implicit val backend: SttpBackend[F, S with WebSockets]
+  implicit val monad: MonadError[F]
+  implicit val convertToFuture: ConvertToFuture[F]
+
+  it should "use pipe to process websocket messages" in {
+    basicRequest
+      .get(uri"$wsEndpoint/ws/send_and_expect_echo")
+      .response(asWebSocketStreamAlways(streams)(functionToPipe {
+        case WebSocketFrame.Text(payload, _, _) =>
+          WebSocketFrame.text(payload + "-echo")
+      }))
+      .send()
+      .map(_ => succeed)
+      .toFuture()
+  }
+
+  def functionToPipe(
+      f: WebSocketFrame.Data[_] => WebSocketFrame
+  ): streams.Pipe[WebSocketFrame.Data[_], WebSocketFrame]
+}
